Refactor: sidebar components
