+++ src/app/App.tsx
@@ -1,6 +1,6 @@
 import { Route, Routes } from 'react-router-dom'
 
-import Sidebar from '../components/Sidebar'
+import Sidebar from '../components/sidebar/Sidebar'
 import Content from '../components/Content'
 
 import Home from './pages/Home'
